Set up the Node structure and made the debug properly functional
Added a node structure to the bvh menu and set up the debug menu using that node structure. Currently the Node is only used to find the level of each box, but we can use it to find the intersection if we use it like a tree structure.
Also currently the debug only shows the chosen layer of boxes, it is easy to switch it to show everything leading up to that point by changing the == to >= in debugdraw